chrome: updated to use init.js.
Instead of loading the stylesheets and scripts directly, it makes
sense to use an init script which can dynamically load them as
appropriate.

For example, if the user requires the right-to-left stylesheet, it
can be loaded instead of the default left-to-right one. We can also
load scripts asynchronously from within init.js -- improving the
perceived load time a little.

And, finally, the `<body>` tag had to be added to the chrome to work
around browsers like IE which automatically put all `<script>`s into
the `<head>` if no `<body>` is present.